condaenv.k_za5rh2.requirements.txt
aiofiles==23.2.1
aiolimiter==1.1.0
asgiref==3.8.1
backoff==2.2.1
bce-python-sdk==0.9.17
bcrypt==4.1.3
beautifulsoup4==4.12.3
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
chroma-hnswlib==0.7.5
chromadb==0.5.4
clevercsv==0.8.2
click==8.1.7
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
deprecated==1.2.14
dill==0.3.8
distro==1.9.0
emoji==2.10.1
exceptiongroup==1.2.1
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
google-auth==2.32.0
googleapis-common-protos==1.63.2
gradio-client==0.11.0
greenlet==3.0.3
grpcio==1.65.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.6
ijson==3.3.0
importlib-metadata==7.1.0
importlib-resources==6.4.0
ipython==8.22.2
jinja2==3.1.4
joblib==1.3.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kubernetes==30.1.0
langchain-community==0.2.7
langchain-huggingface==0.0.3
langdetect==1.0.9
lxml==5.1.0
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
marshmallow==3.20.2
matplotlib==3.9.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.13.3
opencv-python==4.10.0.84
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
overrides==7.7.0
pandas==2.2.2
pillow==10.4.0
pip==24.1.2
posthog==3.5.0
protobuf==4.25.3
pulsar-client==3.5.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pyclipper==1.3.0.post5
pycryptodome==3.20.0
pydantic==2.6.3
pydantic-core==2.16.3
pydub==0.25.1
pyjwt==2.8.0
pymupdf==1.24.0
pymupdfb==1.24.0
pyparsing==3.1.2
pypika==0.48.9
pyproject-hooks==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
rapidfuzz==3.6.1
rapidocr-onnxruntime==1.3.15
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
ruff==0.5.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
sentence-transformers==3.0.1
shapely==2.0.4
shellingham==1.5.4
soupsieve==2.5
spark-ai-python==0.3.15
stack-data==0.6.3
starlette==0.36.3
sympy==1.13.0
tabulate==0.9.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1
tqdm==4.66.2
transformers==4.42.4
triton==2.3.1
typer==0.12.3
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2024.1
unstructured==0.12.6
unstructured-client==0.18.0
urllib3==1.26.18
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websocket-client==1.7.0
websockets==11.0.3
wrapt==1.16.0
zhipuai==2.0.1